old_versions/old_twitter_versions/v1.5/philobot-hashtag-r2/requirements.txt
tweepy
pillow
schedule
request
regex